requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning==1.9.0
albumentations==1.0.0
pandas==1.4.1
numpy>=1.21.0
opencv-python-headless==4.5.5.64
scikit-learn==1.2.2
scikit-image==0.18.2
scipy>=1.11.0
matplotlib==3.4.2
Pillow>=10.0.0
tqdm==4.61.1
tensorboard>=2.0.0